requirements.txt
PyJWT==2.8.0
SQLAlchemy==2.0.30
pytest==8.3.2
httpx==0.27.0
pytest-cov==5.0.0
requests-mock==1.12.1
psycopg2==2.9.9
fastapi[standard]==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
docker==7.1.0